+++ src/__tests__/configs/acala.yml
@@ -3,7 +3,7 @@ endpoint:
   - wss://acala-rpc.dwellir.com
 mock-signature-host: true
 # block: ${env.ACALA_BLOCK_NUMBER}
-block: 6510220
+# block: 6510220
 db: ./db.sqlite
 runtime-log-level: 5
 
@@ -62,12 +62,22 @@ import-storage:
         - free: 1000000000000000
       -
         -
+          - 246gNkjCexYRsCpdjtVhz35sHjcb21jpqipzT9u4uwKV8iEE
+          - token: DOT
+        - free: 1000000000000000
+      -
+        -
           - 5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY
           - token: LDOT
         - free: 1000000000000000
       -
         -
           - 5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY
+          - liquidCrowdloan: 13
+        - free: 1000000000000000
+      -
+        -
+          - 246gNkjCexYRsCpdjtVhz35sHjcb21jpqipzT9u4uwKV8iEE
           - liquidCrowdloan: 13
         - free: 1000000000000000
   GeneralCouncil:
