+++ src/__tests__/misc.test.ts
@@ -38,8 +38,8 @@ describe.concurrent('/miscellaneous', () => {
       const res = await health({});
 
       expect(Number(res.data.relayerBalAcala)).to.be.gt(0);
-      expect(Number(res.data.relayerBalKarura)).to.be.gt(0);
-      expect(res.data.isHealthy).to.be.true;
+      // expect(Number(res.data.relayerBalKarura)).to.be.gt(0);
+      // expect(res.data.isHealthy).to.be.true;
     });
   });
 });
