+++ src/__tests__/xcm.test.ts
@@ -1,13 +1,11 @@
 import { AcalaJsonRpcProvider } from '@acala-network/eth-providers';
 import { ERC20__factory } from '@certusone/wormhole-sdk/lib/cjs/ethers-contracts';
 import { ROUTER_TOKEN_INFO } from '@acala-network/asset-router/dist/consts';
-import { Wallet } from 'ethers';
 import { describe, expect, it } from 'vitest';
 
-import { ETH_RPC } from '../consts';
+import { ETH_RPC, PARA_ID } from '../consts';
 import {
   TEST_ADDR_RELAYER,
-  TEST_KEY,
 } from './testConsts';
 import {
   VAA_10_DAI_ETH_TO_HYDRA,
@@ -22,10 +20,13 @@ import {
   relayAndRouteBatch,
   routeXcm,
   shouldRouteXcm,
-  transferToRouter,
+  sudoTransferToken,
 } from './testUtils';
 
+
 const DAI_ADDR = ROUTER_TOKEN_INFO.dai.acalaAddr;
+
+const provider = new AcalaJsonRpcProvider(ETH_RPC.LOCAL);
 
 const routeXcmArgs = {
   dest: '0x04010200c91f0100525756d2a8c2bb099f2ac22c669bb5a1e5eaf94f687f7b7c5779b288b05bed75',
@@ -33,8 +34,95 @@ const routeXcmArgs = {
   originAddr: ROUTER_TOKEN_INFO.dai.originAddr,
 };
 
-const provider = new AcalaJsonRpcProvider(ETH_RPC.LOCAL);
-const relayer = new Wallet(TEST_KEY.RELAYER, provider);
+describe.concurrent('/shouldRouteXcm', () => {
+  const dest = '0x03010200a9200100d43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d';
+
+  it('when should route', async () => {
+    const res = await shouldRouteXcm(routeXcmArgs);
+    expect(res).toMatchInlineSnapshot(`
+      {
+        "data": {
+          "routerAddr": "0xc5Dd20eD342CEFeCe9E013D7D1763c398E987F7d",
+          "routerChainId": 12,
+          "shouldRoute": true,
+        },
+      }
+    `);
+  });
+
+  describe('when should not route', () => {
+    it('when missing params', async () => {
+      try {
+        await shouldRouteXcm({
+          destParaId: PARA_ID.BASILISK,
+          originAddr: '0x07865c6e87b9f70255377e024ace6630c1eaa37f',
+        });
+        expect.fail('did not throw an err');
+      } catch (err) {
+        expectError(err, ['dest is a required field'], 400);
+      }
+
+      try {
+        await shouldRouteXcm({
+          dest,
+          originAddr: '0x07865c6e87b9f70255377e024ace6630c1eaa37f',
+        });
+        expect.fail('did not throw an err');
+      } catch (err) {
+        expectError(err, ['destParaId is a required field'], 400);
+      }
+
+      try {
+        await shouldRouteXcm({
+          dest,
+          destParaId: PARA_ID.BASILISK,
+        });
+        expect.fail('did not throw an err');
+      } catch (err) {
+        expectError(err, ['originAddr is a required field'], 400);
+      }
+    });
+
+    it('when bad params', async () => {
+      const validArgs = {
+        dest,
+        destParaId: PARA_ID.BASILISK,
+        originAddr: '0x07865c6e87b9f70255377e024ace6630c1eaa37f',
+      };
+
+      let res = await shouldRouteXcm({
+        ...validArgs,
+        destParaId: 1111,
+      });
+      expect(res).to.deep.eq({
+        data: {
+          shouldRoute: false,
+          msg: 'unsupported dest parachain: 1111',
+        },
+      });
+
+      const unsupportedToken = '0x07865c6e87b9f70255377e024ace6630c1e00000';
+      res = await shouldRouteXcm({
+        ...validArgs,
+        originAddr: unsupportedToken,
+      });
+      expect(res).to.deep.eq({
+        data: {
+          shouldRoute: false,
+          msg: `unsupported token on dest parachin 2090. Token origin address: ${unsupportedToken}`,
+        },
+      });
+
+      // TODO: need to validate dest?
+      // res = await shouldRouteXcm({
+      //   ...validArgs,
+      //   dest: '0xabcd'
+      // });
+      // expect(res.data.shouldRoute).to.equal(false);
+      // expect(res.data.msg).to.contain('unsupported router chainId: 8');
+    });
+  });
+});
 
 describe('/routeXcm', () => {
   it('when should route', async () => {
@@ -42,7 +130,7 @@ describe('/routeXcm', () => {
     const { routerAddr } = res.data;
 
     console.log('transferring to router ...');
-    await transferToRouter(routerAddr, relayer);
+    await sudoTransferToken('0xBbBBa9Ebe50f9456E106e6ef2992179182889999', routerAddr, provider, DAI_ADDR, 1.23);
 
     const dai = ERC20__factory.connect(DAI_ADDR, provider);
     const curBalRelayer = (await dai.balanceOf(TEST_ADDR_RELAYER)).toBigInt();
